pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "google-search-mcp"
version = "0.1.0"
description = "A Model Context Protocol (MCP) server that utilizes Google Custom Search (CSE) to retrieve information from the internet"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Shogo Okamoto" }
]
keywords = ["mcp", "google", "search", "llm", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
google-search-mcp = "google_search_mcp:main"
google-search-tool = "google_search_tool:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "integration: end-to-end tests that call real external APIs",
    "slow: tests that take significant time (network I/O, Playwright, etc.)",
]

[tool.hatch.build.targets.wheel]
packages = ["google_search_mcp", "google_search_tool"]
